Fix AllocaOperation using size instead of count in TestRvsdgs (#1342)
The integer input of an AllocaOperation is called size, but it actually
refers to the number of elements of the given type.

This means it should be 1 to get a single `int`. The fact that an `int`
is 4 bytes is already known from its type.

Even if we write `int a[30]`, we should still use the input 1, since the
type will be `[30 x int]`, and we only want a single array.

At least this is what I have gathered from the LLVM->RVSDG conversion. I
also think it makes sense. @caleridas @phate